% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_aov.R
\name{all_aov}
\alias{all_aov}
\title{全面的方差分析及统计检验函数 (展示优化版)}
\usage{
all_aov(factor, var_name)
}
\arguments{
\item{factor}{字符向量，指定数据框中作为分组变量的列名}

\item{var_name}{输入数据，可以是文件路径、数据框或对象名称}
}
\description{
全面的方差分析及统计检验函数 (展示优化版)
}
