% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_factor.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Batch Convert Columns to Factors}
\usage{
convert_to_factor(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A character vector of column names to be converted.}
}
\value{
A data frame with the specified columns converted to factors.
}
\description{
Efficiently converts multiple specified columns in a data frame to factors.
}
\examples{
\dontrun{
  # Convert "Temp" and "Evol" to factors
  data <- convert_to_factor(data, c("Temp", "Evol"))
}
}
