% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_long.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Convert Wide Format Data to Long Format}
\usage{
convert_to_long(
  data,
  id_column = "SampleID",
  variable_column = "Variable",
  value_column = "Value",
  auto_create_id = TRUE
)
}
\arguments{
\item{data}{A data frame containing the wide format data}

\item{id_column}{Character, the name of the identifier column. Default is "SampleID"}

\item{variable_column}{Character, the name for the column storing original column names. Default is "Variable"}

\item{value_column}{Character, the name for the column storing values. Default is "Value"}

\item{auto_create_id}{Logical, whether to automatically create an ID column if it doesn't exist. Default is TRUE}
}
\value{
A data frame in long format
}
\description{
This function converts data from wide format to long format using tidyr::pivot_longer.
}
\examples{
# Create sample data
sample_data <- data.frame(
  SampleID = c("A", "B", "C"),
  Height = c(170, 165, 180),
  Weight = c(65, 60, 75),
  Age = c(25, 30, 28)
)

# Basic usage
long_data <- convert_to_long(sample_data)

# Custom column names
long_data_custom <- convert_to_long(
  sample_data,
  variable_column = "Measurement",
  value_column = "Result"
)

}
