% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{read_clipboard}
\alias{read_clipboard}
\title{Read Data from Clipboard}
\usage{
read_clipboard(sep = "\\t", header = TRUE, ...)
}
\arguments{
\item{sep}{Character string specifying the field separator. Default is "\\t" (tab).}

\item{header}{Logical value indicating whether the data has a header row. Default is TRUE.}

\item{...}{Additional arguments passed to \code{\link[utils]{read.table}}.}
}
\value{
A data frame containing the data from the clipboard.
}
\description{
This function reads tabular data from the system clipboard into a data frame.
It is particularly useful for quickly importing data copied from spreadsheets
or other tabular sources.
}
\details{
This function is a wrapper around \code{read.table("clipboard", ...)} and is
primarily useful in interactive R sessions. It may not work in all environments,
particularly in non-interactive sessions or on some operating systems.
}
\note{
- On Windows, this function uses the system clipboard
- On macOS, it uses the pbpaste command
- On Linux, it may require xclip or xsel to be installed
}
\examples{
\dontrun{
# Copy some tabular data to your clipboard (e.g., from Excel)
# Then run:
my_data <- read_clipboard()

# If your data uses a different separator (e.g., comma):
my_data <- read_clipboard(sep = ",")

# If your data doesn't have a header:
my_data <- read_clipboard(header = FALSE)
}

}
\seealso{
\code{\link[utils]{read.table}} for the underlying function used to read the data
\code{\link[clipr]{read_clip}} for a more robust clipboard reading function
}
