% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pairwise.R
\name{split_pairwise}
\alias{split_pairwise}
\title{Generate Pairwise Subsets from a Data Frame}
\usage{
split_pairwise(data, group_col)
}
\arguments{
\item{data}{A data frame.}

\item{group_col}{Character string. The name of the grouping column.}
}
\value{
A named list of data frames. Names are formatted as "Group1_vs_Group2".
}
\description{
Creates a list of data frames, where each data frame contains data for
only two specific levels of the grouping variable. It generates all possible
pairwise combinations.
}
